re-organize code.
