+++ src/components/ArtistCard/ArtistCard.tsx
